exp/clients/horizon: stream ledgers (#1066)
* stream ledgers

* test for stream error

* add comment
