+++ exp/clients/horizon/client.go
@@ -1,12 +1,18 @@
 package horizonclient
 
 import (
+	"bufio"
+	"bytes"
 	"context"
 	"encoding/json"
+	"fmt"
+	"io"
 	"net/http"
+	"net/url"
 	"strings"
 	"time"
 
+	"github.com/manucorporat/sse"
 	hProtocol "github.com/stellar/go/protocols/horizon"
 	"github.com/stellar/go/protocols/horizon/operations"
 	"github.com/stellar/go/support/app"
@@ -50,6 +56,135 @@ func (c *Client) sendRequest(hr HorizonRequest, a interface{}) (err error) {
 	err = decodeResponse(resp, &a)
 	cancel()
 	return
+}
+
+// stream handles connections to endpoints that support streaming on an horizon server
+func (c *Client) stream(
+	ctx context.Context,
+	streamURL string,
+	handler func(data []byte) error,
+) error {
+	su, err := url.Parse(streamURL)
+	if err != nil {
+		return errors.Wrap(err, "Error parsing stream url")
+	}
+
+	query := su.Query()
+	if query.Get("cursor") == "" {
+		query.Set("cursor", "now")
+	}
+
+	for {
+		// updates the url with new cursor
+		su.RawQuery = query.Encode()
+		req, err := http.NewRequest("GET", fmt.Sprintf("%s", su), nil)
+		if err != nil {
+			return errors.Wrap(err, "Error creating HTTP request")
+		}
+		req.Header.Set("Accept", "text/event-stream")
+		// to do: confirm name and version
+		req.Header.Set("X-Client-Name", "go-stellar-sdk")
+		req.Header.Set("X-Client-Version", app.Version())
+
+		// We can use c.HTTP here because we set Timeout per request not on the client. See sendRequest()
+		resp, err := c.HTTP.Do(req)
+		if err != nil {
+			return errors.Wrap(err, "Error sending HTTP request")
+		}
+
+		// Expected statusCode are 200-299
+		if !(resp.StatusCode >= 200 && resp.StatusCode < 300) {
+			return fmt.Errorf("Got bad HTTP status code %d", resp.StatusCode)
+		}
+		defer resp.Body.Close()
+
+		reader := bufio.NewReader(resp.Body)
+
+		// Read events one by one. Break this loop when there is no more data to be
+		// read from resp.Body (io.EOF).
+	Events:
+		for {
+			// Read until empty line = event delimiter. The perfect solution would be to read
+			// as many bytes as possible and forward them to sse.Decode. However this
+			// requires much more complicated code.
+			// We could also write our own `sse` package that works fine with streams directly
+			// (github.com/manucorporat/sse is just using io/ioutils.ReadAll).
+			var buffer bytes.Buffer
+			nonEmptylinesRead := 0
+			for {
+				// Check if ctx is not cancelled
+				select {
+				case <-ctx.Done():
+					return nil
+				default:
+					// Continue
+				}
+
+				line, err := reader.ReadString('\n')
+				if err != nil {
+					if err == io.EOF || err == io.ErrUnexpectedEOF {
+						// We catch EOF errors to handle two possible situations:
+						// - The last line before closing the stream was not empty. This should never
+						//   happen in Horizon as it always sends an empty line after each event.
+						// - The stream was closed by the server/proxy because the connection was idle.
+						//
+						// In the former case, that (again) should never happen in Horizon, we need to
+						// check if there are any events we need to decode. We do this in the `if`
+						// statement below just in case if Horizon behaviour changes in a future.
+						//
+						// From spec:
+						// > Once the end of the file is reached, the user agent must dispatch the
+						// > event one final time, as defined below.
+						if nonEmptylinesRead == 0 {
+							break Events
+						}
+					} else {
+						return errors.Wrap(err, "Error reading line")
+					}
+				}
+
+				buffer.WriteString(line)
+
+				if strings.TrimRight(line, "\n\r") == "" {
+					break
+				}
+
+				nonEmptylinesRead++
+			}
+
+			events, err := sse.Decode(strings.NewReader(buffer.String()))
+			if err != nil {
+				return errors.Wrap(err, "Error decoding event")
+			}
+
+			// Right now len(events) should always be 1. This loop will be helpful after writing
+			// new SSE decoder that can handle io.Reader without using ioutils.ReadAll().
+			for _, event := range events {
+				if event.Event != "message" {
+					continue
+				}
+
+				// Update cursor with event ID
+				if event.Id != "" {
+					query.Set("cursor", event.Id)
+				}
+
+				switch data := event.Data.(type) {
+				case string:
+					err = handler([]byte(data))
+					err = errors.Wrap(err, "Handler error")
+				case []byte:
+					err = handler(data)
+					err = errors.Wrap(err, "Handler error")
+				default:
+					err = errors.New("Invalid event.Data type")
+				}
+				if err != nil {
+					return err
+				}
+			}
+		}
+	}
 }
 
 // getHorizonUrl strips all slashes(/) at the end of HorizonURL if any, then adds a single slash
@@ -115,7 +250,7 @@ func (c *Client) Assets(request AssetRequest) (assets hProtocol.AssetsPage, err 
 // Stream is for endpoints that support streaming
 func (c *Client) Stream(ctx context.Context, request StreamRequest, handler func(interface{})) (err error) {
 
-	err = request.Stream(ctx, c.getHorizonURL(), handler)
+	err = request.Stream(ctx, c, handler)
 	return
 }
