+++ exp/clients/horizon/main_test.go
@@ -520,71 +520,6 @@ func TestAssetsRequest(t *testing.T) {
 
 }
 
-func TestLedgerDetail(t *testing.T) {
-	hmock := httptest.NewClient()
-	client := &Client{
-		HorizonURL: "https://localhost/",
-		HTTP:       hmock,
-	}
-
-	// invalid parameters
-	var sequence uint32 = 0
-	hmock.On(
-		"GET",
-		"https://localhost/ledgers/",
-	).ReturnString(200, ledgerResponse)
-
-	_, err := client.LedgerDetail(sequence)
-	// error case: invalid sequence
-	if assert.Error(t, err) {
-		assert.Contains(t, err.Error(), "Invalid sequence number provided")
-	}
-
-	// happy path
-	hmock.On(
-		"GET",
-		"https://localhost/ledgers/69859",
-	).ReturnString(200, ledgerResponse)
-
-	sequence = 69859
-	ledger, err := client.LedgerDetail(sequence)
-	ftc := int32(1)
-
-	if assert.NoError(t, err) {
-		assert.Equal(t, ledger.ID, "71a40c0581d8d7c1158e1d9368024c5f9fd70de17a8d277cdd96781590cc10fb")
-		assert.Equal(t, ledger.PT, "300042120331264")
-		assert.Equal(t, ledger.Sequence, int32(69859))
-		assert.Equal(t, ledger.FailedTransactionCount, &ftc)
-	}
-
-	// failure response
-	hmock.On(
-		"GET",
-		"https://localhost/ledgers/69859",
-	).ReturnString(404, notFoundResponse)
-
-	_, err = client.LedgerDetail(sequence)
-	if assert.Error(t, err) {
-		assert.Contains(t, err.Error(), "Horizon error")
-		horizonError, ok := err.(*Error)
-		assert.Equal(t, ok, true)
-		assert.Equal(t, horizonError.Problem.Title, "Resource Missing")
-	}
-
-	// connection error
-	hmock.On(
-		"GET",
-		"https://localhost/ledgers/69859",
-	).ReturnError("http.Client error")
-
-	_, err = client.LedgerDetail(sequence)
-	if assert.Error(t, err) {
-		assert.Contains(t, err.Error(), "http.Client error")
-		_, ok := err.(*Error)
-		assert.Equal(t, ok, false)
-	}
-}
-
 func TestMetrics(t *testing.T) {
 	hmock := httptest.NewClient()
 	client := &Client{
